Add files via upload
By taking Super Market Sales dataset ,I have made a prediction model using XgBoost algorithm of machine learning which predicts
the sales of different items available in the departmental store like which is the maximum selling product in the store and which is
the minimum , so that the store managers can maintain the quantity of those products accordingly and they may also know what will
the challenges they will be going to face in future. This model helps in the overall growth and productivity of the store and enhance
their sales.